+++ python/megane/parsers/__init__.py
@@ -8,6 +8,7 @@ from megane.parsers.mol2 import load_mol2
 from megane.parsers.netcdf import load_netcdf
 from megane.parsers.pdb import load_pdb
 from megane.parsers.traj import load_traj
+from megane.parsers.vasp import load_vasp
 from megane.parsers.xtc import load_trajectory
 from megane.parsers.xyz import load_xyz_trajectory
 
@@ -25,5 +26,6 @@ __all__ = [
     "load_sdf",
     "load_traj",
     "load_trajectory",
+    "load_vasp",
     "load_xyz_trajectory",
 ]